+++ src/youtube_automation/domains/distrokid/metadata.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 import re
 from pathlib import Path
 
-from youtube_automation.core.adapters.errors import ConfigError
+from youtube_automation.core.errors import ConfigError
 
 # アルバム情報表の `項目` ラベル → payload canonical key。
 _ALBUM_LABEL_MAP = {
